Added the Testing of getValues()
Bug found for getting Fahrenheit values : during the conversion to
Kelvins the value in kelvins was not a round number, and therefore when
we converted back the exact number was not found. This was because we
didn't have a maximum precision.
We are going to correct this bug based on what the professor wrote in
myCourses, which is that the number should be precise up to 1E-6