Fixed the Rounding bug in the Temperature class
We now round the numbers to 1e-6 before returning it. We only do this in
the convertFromKelvin() method because in the convertToKelvin() method
we want to keep the number as precise as possible, since this is the
number that will be stored but will never be seen by the user (since it
will be rounded back in the convertFromKelvin() method).